fixed handler, fixed errors, removed ponters in business logic for replay servce

+++ internal/pkg/apperror/error.go
@@ -3,11 +3,11 @@ package apperror
 var _ error = &Err{}
 
 type Err struct {
-	Code uint16
+	Code int
 	Msg  string
 }
 
-func New(code uint16, msg string) Err {
+func New(code int, msg string) Err {
 	return Err{
 		Code: code,
 		Msg:  msg,
